+++ app/assets/stylesheets/application.css
@@ -240,15 +240,28 @@ body{
   font-weight: bold;
 }
 
-#player-buttons {
+.player-buttons {
   border-radius: 100%;
+  border-color: black;
   padding: 10px;
-  background-color: #6E6662;
+  background-color:rgba(0, 0, 0, 0.7);
+  opacity: 0.9;
   color: white;
-  height: 50px;
-  margin: 10px;
-  font-size: .8em;
+  margin: 10px 0 10px 40px;
+  font-size: 2em;
   font-weight: normal;
+}
+
+.player-buttons:hover{
+  background-color: black;
+  filter: grayscale(0%);
+  opacity: 1;
+}
+
+.player-buttons:focus{
+  background-color: black;
+  filter: grayscale(0%);
+  opacity: 1;
 }
 
 #new-pack-form{
