+++ app/javascript/packs/application.js
@@ -7,8 +7,12 @@ import RedBox from 'redbox-react'
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faPlayCircle } from '@fortawesome/free-solid-svg-icons'
+import { faPause } from '@fortawesome/free-solid-svg-icons'
+import { faDownload } from '@fortawesome/free-solid-svg-icons'
 
 library.add(faPlayCircle)
+library.add(faPause)
+library.add(faDownload)
 
 document.addEventListener('DOMContentLoaded', () => {
   let reactElement = document.getElementById('app')
