+++ app/javascript/react/containers/CollectionShowContainer.js
@@ -1,8 +1,8 @@
 import React, { Component } from "react"
 import { Link } from "react-router"
-import SamplesContainer from "./SamplesContainer"
-import SampleTile from "./SampleTile"
-import SamplesFormContainer from "./SamplesFormContainer"
+import SamplesContainer from "../containers/SamplesContainer"
+import SampleTile from "../components/SampleTile"
+import SamplesFormContainer from "../containers/SamplesFormContainer"
 
 class CollectionShowContainer extends Component {
   constructor(props) {
@@ -10,10 +10,9 @@ class CollectionShowContainer extends Component {
     this.state = {
       collection: {},
       creator: "",
-      samples: {},
+      samples: [],
       currentUser: ""
     }
-    this.forceRender = this.forceRender.bind(this)
   }
 
   componentDidMount() {
@@ -21,7 +20,6 @@ class CollectionShowContainer extends Component {
 
     fetch(`/api/v1/collections/${collectionId}`, {
       credentials: "same-origin",
-      method: "GET",
       headers: {
         Accept: "application/json",
         "Content-Type": "application/json"
@@ -42,10 +40,10 @@ class CollectionShowContainer extends Component {
         id: 0,
         role: "Not Signed In"
       };
-      if (body.current_user === null) {
-        body.current_user = noUser;
+      if (body.collection.currentUser === null) {
+        body.collection.currentUser = noUser;
       }
-      this.setState({ collection: body.collection, creator: body.creator, currentUser: body.current_user, samples: body.samples })
+      this.setState({ collection: body.collection, creator: body.collection.user_id, currentUser: body.collection.currentUser, samples: body.collection.samples })
       return this.state
     })
     .then(stateFul => {
@@ -54,8 +52,26 @@ class CollectionShowContainer extends Component {
     .catch(error => console.error(`Error in fetch: ${error.message}`))
   }
 
-  forceRender() {
-    setTimeout(function(){this.componentDidMount()}, 3000)
+  onSubmit(payload) {
+    fetch(`/api/v1/samples/${this.collectionId}/samples`, {
+      credentials: "same-origin",
+      method: "POST",
+      body: JSON.stringify(payload)
+    })
+    .then(response => {
+      if (response.ok) {
+        return response;
+      } else {
+        let errorMessage = `${response.status} (${response.statusText})`,
+          error = new Error(errorMessage);
+        throw error;
+      }
+    })
+    .then(response => response.json())
+    .then(body => {
+      this.setState({ samples: this.state.samples.concat(body.sample) });
+    })
+    .catch(error => console.error(`Error in fetch: ${error.message}`));
   }
 
   render() {
@@ -64,8 +80,9 @@ class CollectionShowContainer extends Component {
 
     if (this.state.samples.length >= 0){
         samples = <SamplesContainer
-          currentUserId={this.state.currentUser.id}
-          collectionId={this.props.params.id}
+          currentUserId={currentUserId}
+          collectionId={this.state.collection.id}
+          samples={this.state.samples}
         />
     }
     return (
@@ -91,10 +108,9 @@ class CollectionShowContainer extends Component {
           <SamplesFormContainer
             collectionId={this.props.params.id}
             userId={this.state.currentUser.id}
-            addSample={this.addSample}
-            force={this.forceRender}
             creatorId={this.state.collection.user_id}
             currentUser={this.state.currentUser}
+            onSubmit={this.onSubmit}
           />
         </div>
       </div>
