+++ app/serializers/collection_serializer.rb
@@ -1,0 +1,10 @@
+class CollectionSerializer < ActiveModel::Serializer
+  attributes :id, :name, :img, :description, :user_id, :currentUser
+
+  def currentUser
+    return current_user
+  end
+
+  belongs_to :user
+  has_many :samples
+end
