+++ config/initializers/wrap_parameters.rb
@@ -2,6 +2,7 @@
 
 # This file contains settings for ActionController::ParamsWrapper which
 # is enabled by default.
+ActiveModelSerializers.config.adapter = :json
 
 # Enable parameter wrapping for JSON. You can disable this by setting :format to an empty array.
 ActiveSupport.on_load(:action_controller) do
