+++ package.json
@@ -25,7 +25,8 @@
     "fetch-ponyfill": "^6.0.2",
     "jasmine-core": "~2.4.1",
     "jasmine-enzyme": "~3.4.0",
-    "karma": "~0.13.22",
+    "karma": "^4.1.0",
+    "karma-chrome-launcher": "^2.2.0",
     "karma-coverage": "0.5.5",
     "karma-jasmine": "~0.3.8",
     "karma-phantomjs-launcher": "~1.0.4",
